moved get_all_items to inventory
